Extract Markdown autocomplete into dedicated modules
Move the slash-command and code-fence completion logic out of Markdown.tsx
(919 -> 573 lines) into markdown_completions.ts, and the shared editor
helpers (insertText/replaceSelection/uploadImageAndInsert) into
markdown_editor_utils.ts to avoid a circular import. Tests move alongside
into markdown_completions.spec.ts. No behavior change.
